Rc/v2.1.0

+++ feature-account-api/src/main/java/io/novafoundation/nova/feature_account_api/domain/model/MetaAccount.kt
@@ -89,7 +89,7 @@ fun MetaAccount.addressIn(chain: Chain): String? {
     return when {
         hasChainAccountIn(chain.id) -> chain.addressOf(chainAccounts.getValue(chain.id).accountId)
         chain.isEthereumBased -> ethereumAddress?.ethereumAddressToHex()
-        else -> substrateAccountId.toAddress(chain.addressPrefix.toByte())
+        else -> substrateAccountId.toAddress(chain.addressPrefix.toShort())
     }
 }
